requirements.txt
# 肺炎识别系统依赖
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
streamlit>=1.0.0
pillow>=8.0.0
scikit-learn>=0.24.0
PyQt5>=5.15.0
scipy>=1.10.0